Actualiza nombres de paquetes para cumplir con convenciones de Java
Este cambio cambia el arbol del proyecto completo. Implica cambios en los mappers.
