+++ src/main/java/edu/eci/pdsw/epicwino/logica/dao/PersistenceException.java
@@ -1,0 +1,21 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package edu.eci.pdsw.epicwino.logica.dao;
+
+/**
+ *
+ * @author Esteban
+ */
+public class PersistenceException extends Exception {
+    
+    public PersistenceException(String message) {
+        super(message);
+    }
+
+    public PersistenceException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
