+++ src/main/java/edu/eci/pdsw/epicwino/logica/dao/RecursoDAO.java
@@ -1,0 +1,22 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package edu.eci.pdsw.epicwino.logica.dao;
+
+import edu.eci.pdsw.epicwino.logica.entidades.Recurso;
+import java.util.List;
+
+/**
+ *
+ * @author Esteban
+ */
+public interface RecursoDAO {
+    
+    public void save(Recurso rec) throws PersistenceException;
+    
+    public Recurso load(int idRec) throws PersistenceException;
+    
+     public List<Recurso> loadRecursos() throws PersistenceException;
+}
